requirements.txt
torch>=1.6.1
torchvision>=0.10.1
numpy
pandas
tqdm
tensorboardX>=1.14
scipy 
opencv-python
clean-fid
pytorch-lightning==1.5.6
PyYAML==6.0
tensorboard>=2.2.0
torchsummary==1.5.1
